Increase coverage for generate-config
- Move interactive methods to interactive.go and ignore in coverage
- Add test to run GenerateConfig function
